pyproject.toml
[project]
name = "veo3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.36.0",
    "pillow>=11.3.0",
    "pydantic-ai>=1.2.1",
    "python-dotenv>=1.1.1",
]
